requirements.txt
grad_cam==1.4.8
matplotlib==3.8.2
numpy==2.1.3
openai==0.28.0
opencv_python==4.9.0.80
pandas==2.2.3
Pillow==11.0.0
python-dotenv==1.0.1
Requests==2.32.3
timm==0.9.12
torch==2.2.0
torchvision==0.17.0
transformers==4.30.2
